+++ src/library/DIPS.Mobile.UI/Components/Pages/ContentPage.cs
@@ -1,3 +1,4 @@
+using DIPS.Mobile.UI.Components.Navigation.FloatingNavigationButton;
 using Colors = DIPS.Mobile.UI.Resources.Colors.Colors;
 
 namespace DIPS.Mobile.UI.Components.Pages
@@ -5,6 +6,7 @@ namespace DIPS.Mobile.UI.Components.Pages
     public partial class ContentPage : Microsoft.Maui.Controls.ContentPage
     {
         public static readonly ColorName BackgroundColorName = ColorName.color_neutral_10;
+        private bool m_hasAppeared;
 
         public ContentPage()
         {
@@ -16,6 +18,30 @@ namespace DIPS.Mobile.UI.Components.Pages
             SetColors(Application.Current.RequestedTheme);
             Application.Current.RequestedThemeChanged +=
                 OnRequestedThemeChanged; //Can not use AppThemeBindings because that makes the navigation page bar background flash on Android, so we listen to changes and set the color our self
+        }
+
+        protected override void OnAppearing()
+        {
+            base.OnAppearing();
+
+            m_hasAppeared = true;
+
+            HideOrShowFloatingNavigationMenu();
+        }
+
+        private void HideOrShowFloatingNavigationMenu()
+        {
+            if (!m_hasAppeared)
+                return;
+            
+            if (ShouldHideFloatingNavigationMenuButton)
+            {
+                FloatingNavigationButtonService.Hide();
+            }
+            else
+            {
+                FloatingNavigationButtonService.Show();
+            }
         }
 
         private void SetColors(AppTheme osAppTheme)
@@ -36,12 +62,14 @@ namespace DIPS.Mobile.UI.Components.Pages
         protected override void OnDisappearing()
         {
             base.OnDisappearing();
+
+            m_hasAppeared = false;
+            
             if (Application.Current != null)
             {
                 Application.Current.RequestedThemeChanged -= OnRequestedThemeChanged;
             }
 
         }
-
     }
 }